SchemeStealer/requirements.txt
streamlit==1.31.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
scikit-learn==1.4.0
scikit-image==0.22.0
rembg==2.0.56
pillow==10.2.0
scipy==1.12.0